Migrate from Retrofit to Ktor

+++ core/data/src/main/java/com/anilyilmaz/awesomesunsetwallpapers/core/data/repository/PhotoRepository.kt
@@ -5,7 +5,7 @@ import com.anilyilmaz.awesomesunsetwallpapers.core.network.model.PexelsPhoto
 import kotlinx.coroutines.flow.Flow
 
 interface PhotoRepository {
-    suspend fun getPhoto(id: Int): PexelsPhoto
+    suspend fun getPhoto(id: Long): PexelsPhoto
 
     fun getPhotosWithQuery(query: List<String>): Flow<PagingData<PexelsPhoto>>
 }
